SimonStewart: Removing the selenium rc reference from the webdriver servlet. The grid build fails in the same way as on a pristine client, so I assume that this change is a no-op.
r16527

+++ java/server/src/org/openqa/selenium/remote/server/DefaultSession.java
@@ -128,7 +128,7 @@ public class DefaultSession implements Session {
     lastAccess = System.currentTimeMillis();
   }
 
-  public boolean isTimedOut(int timeout) {
+  public boolean isTimedOut(long timeout) {
     return timeout > 0 && (lastAccess + timeout) < System.currentTimeMillis();
   }
