+++ java/server/test/org/openqa/selenium/remote/server/testing/TestSession.java
@@ -91,7 +91,7 @@ public class TestSession implements Session {
     return null;
   }
 
-  public boolean isTimedOut(int timeout) {
+  public boolean isTimedOut(long timeout) {
     return timeout > 0 && (lastAccess + timeout) < System.currentTimeMillis();
   }
